Common implementation for GetExpressionType
